DP-13: Refactor date picker components to use consistent casing and restructure imports; update color constants for improved clarity
