+++ shared/components/datePicker/components/datePickerModal/datePickerInput.tsx
@@ -3,9 +3,9 @@ import { StyleSheet, TouchableOpacity } from "react-native";
 import Icon from "react-native-vector-icons/MaterialIcons";
 
 import { Text } from "@/shared/components/themed/text";
-import { Colors } from "@/shared/constants/colors";
+import { Colors } from "@/shared/constants/Colors";
 
-import { IDatePickerState } from "../interfaces";
+import { IDatePickerState } from "../../interfaces";
 
 interface IDatePickerInput {
   setState: (value: React.SetStateAction<IDatePickerState>) => void;
