+++ shared/components/datePicker/components/datePickerModal/index.tsx
@@ -2,9 +2,9 @@ import React, { FC } from "react";
 import { Modal, StyleSheet } from "react-native";
 
 import { View } from "@/shared/components/themed/view";
-import { Colors } from "@/shared/constants/colors";
+import { Colors } from "@/shared/constants/Colors";
 
-import { IDatePickerState } from "../interfaces";
+import { IDatePickerState } from "../../interfaces";
 import { DatePickerFooter } from "./datePickerFooter";
 import { DatePickerHeader } from "./datePickerHeader";
 import { DatePickerMain } from "./datePickerMain";
